refactor(di): Improve DI registration behavior
The service provider should not be created until it is final, larger apps may have a lot of registrations, and it can have a major negative impact on startup performance.

Instead, allow the authentication scheme to be specified as a parameter, with a default, and remove from configuration object.

Additionally, use the options pattern for FishyFlipOptions instead of registering a singleton object. (Registered as singleton for now, will follow up with conversion to IOptions<T>)