Proposal: Better dependency injection behavior (#310)
* refactor: Use `.Replace` instead of removing and adding.

* feat(configuration): Introduce constants for various configuration defaults

* refactor(di): Improve DI registration behavior

The service provider should not be created until it is final, larger apps may have a lot of registrations, and it can have a major negative impact on startup performance.

Instead, allow the authentication scheme to be specified as a parameter, with a default, and remove from configuration object.

Additionally, use the options pattern for FishyFlipOptions instead of registering a singleton object. (Registered as singleton for now, will follow up with conversion to IOptions<T>)

* doc(configuration): Mark authentication scheme properties as obsolete

This was done instead of removal, so that anyone using this gets a hint on where to look for the correct way to do this.

* refactor(options): Use the options pattern for FishyFlipOptions

* refactor: Remove options proxy registration

* fix: Add missing parameter to typed fishy flip method

* refactor: Use generic add fishyflip method for default registration

By flipping these implementations, we avoid the need to remove the registration and add a new one. Further, by not using TryAddScoped, we allow replacement using the add method since the last registration is what will be used.

* fix: Pass configuration section to configure method

* feat: Validate configuration using data annotations

Heavy handed for a single property, but allows future additions here too.